Add missing adapter data to Bluetooth config entry titles
For older config entries, the title would be the mac address
which was confusing for users with multiple adapters or if
they did not know the mac and model of the adapter. Update
the title to reflect the adapter title details if the title
is the mac address
