+++ homeassistant/components/bluetooth/util.py
@@ -2,7 +2,14 @@
 
 from __future__ import annotations
 
-from bluetooth_adapters import BluetoothAdapters
+from bluetooth_adapters import (
+    ADAPTER_ADDRESS,
+    ADAPTER_MANUFACTURER,
+    AdapterDetails,
+    BluetoothAdapters,
+    adapter_model,
+    adapter_unique_name,
+)
 from bluetooth_data_tools import monotonic_time_coarse
 
 from homeassistant.core import callback
@@ -69,3 +76,12 @@ def async_load_history_from_system(
                 connectable_loaded_history[address] = service_info
 
     return all_loaded_history, connectable_loaded_history
+
+
+@callback
+def adapter_title(adapter: str, details: AdapterDetails) -> str:
+    """Return the adapter title."""
+    unique_name = adapter_unique_name(adapter, details[ADAPTER_ADDRESS])
+    model = adapter_model(details)
+    manufacturer = details[ADAPTER_MANUFACTURER] or "Unknown"
+    return f"{manufacturer} {model} ({unique_name})"
